test(switchyard): relocate component tests
